# src/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "UMC 7th",
    "description": "UMC 7th Node.js 테스트 프로젝트입니다.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/reviews/user/{userId}": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "사용자별 리뷰 목록 조회 API",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "사용자 ID"
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 리뷰 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "reviewId": {
                            "type": "number"
                          },
                          "userId": {
                            "type": "number"
                          },
                          "storeId": {
                            "type": "number"
                          },
                          "missionId": {
                            "type": "number",
                            "nullable": true
                          },
                          "rating": {
                            "type": "number"
                          },
                          "reviewText": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "isVerified": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string",
                          "example": "USER_NOT_FOUND"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 404
                        },
                        "reason": {
                          "type": "string",
                          "example": "해당 사용자를 찾을 수 없습니다."
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/missions/store/{storeId}": {
      "get": {
        "tags": [
          "Missions"
        ],
        "summary": "사용자별 미션 목록 조회 API",
        "description": "",
        "parameters": [
          {
            "name": "storeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "사용자 ID",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 미션 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "missionId": {
                            "type": "number"
                          },
                          "storeId": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          },
                          "points": {
                            "type": "number"
                          },
                          "startDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endDate": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "사용자를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string",
                          "example": "USER_NOT_FOUND"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 404
                        },
                        "reason": {
                          "type": "string",
                          "example": "해당 사용자를 찾을 수 없습니다."
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/missions/user/{userId}": {
      "get": {
        "tags": [
          "Missions"
        ],
        "summary": "가게별 미션 목록 조회 API",
        "description": "",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "storeId",
            "in": "path",
            "description": "가게 ID",
            "required": true,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "in": "query",
            "description": "활성 미션만 조회 (optional)",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "가게 미션 목록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "missionId": {
                            "type": "number"
                          },
                          "storeId": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          },
                          "points": {
                            "type": "number"
                          },
                          "startDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "endDate": {
                            "type": "string",
                            "format": "date-time",
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "가게를 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string",
                          "example": "STORE_NOT_FOUND"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 404
                        },
                        "reason": {
                          "type": "string",
                          "example": "해당 가게를 찾을 수 없습니다."
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/signup": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "회원 가입 API",
        "description": "",
        "responses": {
          "200": {
            "description": "회원 가입 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "preferCategory": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "description": "선호하는 음식 카테고리 이름 목록"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "회원 가입 실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string",
                          "example": "INVALID_INPUT"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 400
                        },
                        "reason": {
                          "type": "string",
                          "example": "잘못된 입력값입니다."
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "이미 존재하는 이메일",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string",
                          "example": "EMAIL_ALREADY_EXISTS"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 409
                        },
                        "reason": {
                          "type": "string",
                          "example": "이미 존재하는 이메일입니다."
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "name": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string",
                    "enum": [
                      "MALE",
                      "FEMALE",
                      "OTHER"
                    ]
                  },
                  "username": {
                    "type": "string"
                  },
                  "birth": {
                    "type": "string",
                    "format": "date"
                  },
                  "address": {
                    "type": "string"
                  },
                  "detailAddress": {
                    "type": "string"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "pattern": "^\\d{2,3}-\\d{3,4}-\\d{4}$"
                  },
                  "preferences": {
                    "type": "array",
                    "items": {
                      "type": "number",
                      "description": "음식 카테고리 ID"
                    }
                  }
                },
                "required": [
                  "email",
                  "name",
                  "gender",
                  "username",
                  "birth",
                  "phoneNumber",
                  "preferences"
                ]
              }
            }
          }
        }
      }
    },
    "/reviews": {
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "리뷰 작성 API",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "리뷰 작성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "object",
                      "properties": {
                        "reviewId": {
                          "type": "number"
                        },
                        "userId": {
                          "type": "number"
                        },
                        "storeId": {
                          "type": "number"
                        },
                        "missionId": {
                          "type": "number",
                          "nullable": true
                        },
                        "rating": {
                          "type": "number"
                        },
                        "reviewText": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "isVerified": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number"
                  },
                  "storeId": {
                    "type": "number"
                  },
                  "missionId": {
                    "type": "number",
                    "nullable": true
                  },
                  "rating": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 5
                  },
                  "reviewText": {
                    "type": "string"
                  }
                },
                "required": [
                  "userId",
                  "storeId",
                  "rating",
                  "reviewText"
                ]
              }
            }
          }
        }
      }
    },
    "/missions": {
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "미션 생성 API",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "미션 생성 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "object",
                      "properties": {
                        "missionId": {
                          "type": "number"
                        },
                        "storeId": {
                          "type": "number"
                        },
                        "description": {
                          "type": "string"
                        },
                        "points": {
                          "type": "number"
                        },
                        "startDate": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "endDate": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "미션 생성 실패",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "FAIL"
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "errorCode": {
                          "type": "string"
                        },
                        "statusCode": {
                          "type": "number",
                          "example": 400
                        },
                        "reason": {
                          "type": "string"
                        },
                        "data": {
                          "type": "object",
                          "nullable": true
                        }
                      }
                    },
                    "success": {
                      "type": "object",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "storeId": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "points": {
                    "type": "number"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time",
                    "nullable": true
                  }
                },
                "required": [
                  "storeId",
                  "description",
                  "points"
                ]
              }
            }
          }
        }
      }
    },
    "/user-missions": {
      "post": {
        "tags": [
          "Missions"
        ],
        "summary": "사용자 미션 참여 API",
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "미션 참여 성공",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "resultType": {
                      "type": "string",
                      "example": "SUCCESS"
                    },
                    "error": {
                      "type": "object",
                      "nullable": true
                    },
                    "success": {
                      "type": "object",
                      "properties": {
                        "userMissionId": {
                          "type": "number"
                        },
                        "userId": {
                          "type": "number"
                        },
                        "missionId": {
                          "type": "number"
                        },
                        "status": {
                          "type": "string",
                          "enum": [
                            "STARTED",
                            "COMPLETED",
                            "FAILED"
                          ]
                        },
                        "startedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "completedAt": {
                          "type": "string",
                          "format": "date-time",
                          "nullable": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number"
                  },
                  "missionId": {
                    "type": "number"
                  }
                },
                "required": [
                  "userId",
                  "missionId"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}